Added test data into the root directory named movies.csv, Broke the tests into different files. Fixed typo in ui 11/05/25
