+++ App/MainWindow.xaml
@@ -67,7 +67,7 @@
             <TextBlock Text="Year:" Grid.Row="4" Grid.Column="0" Margin="10"/>
             <TextBox Text="{Binding NewYear, UpdateSourceTrigger=PropertyChanged}" Grid.Row="4" Grid.Column="1" Margin="10"/>
 
-            <Button Content="Add Book" Command="{Binding AddMovieCommand}" Margin="10" Grid.Row="5"/>
+            <Button Content="Add Movie" Command="{Binding AddMovieCommand}" Margin="10" Grid.Row="5"/>
 
             <TextBlock Text="Search:" Grid.Row="6" Grid.Column="0" Margin="10"/>
             <TextBox x:Name="TitleSearchBox"  Margin="10" Grid.Row="6" Grid.Column="1"/>
